requirements.txt
ansible-core>=2.17.0
ansible-lint>=24.0.0
yamllint>=1.35.0
molecule>=24.0.0
molecule-plugins[docker]>=23.5.0
docker>=7.0.0
minio>=7.1.4
